+++ src/web/mod.rs
@@ -37,6 +37,7 @@ mod apache;
 mod caddy;
 mod cisco_asa;
 mod firewall;
+mod general;
 mod geomap;
 mod haproxy;
 mod nginx;
@@ -75,6 +76,7 @@ pub async fn serve(state: Arc<AppState>) -> anyhow::Result<()> {
         .route("/web/haproxy", get(haproxy::dashboard))
         .route("/firewall/cisco_asa", get(cisco_asa::dashboard))
         .route("/firewall/panos", get(panos::dashboard))
+        .route("/general/syslog", get(general::dashboard))
         .route("/apache", get(|uri: Uri| moved(uri, "/web/apache")))
         .route("/apache/recent", get(|uri: Uri| moved(uri, "/web/apache/recent")))
         .route("/nginx", get(|uri: Uri| moved(uri, "/web/nginx")))